Logging functions are now available to C-Users, too. Cleaning up headers. Changing spelling of some enums to camel. Ply loader loads tristrips correctly now.
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@425 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ test/unit/utTriangulate.cpp
@@ -38,8 +38,8 @@ void TriangulateProcessTest :: setUp (void)
 			// construct fully convex input data in ccw winding, xy plane
 			aiVector3D& v = pcMesh->mVertices[pcMesh->mNumVertices++];
 			v.z = 0.f;
-			v.x = cos (p * AI_MATH_TWO_PI/face.mNumIndices);
-			v.y = sin (p * AI_MATH_TWO_PI/face.mNumIndices);
+			v.x = cos (p * (float)(AI_MATH_TWO_PI)/face.mNumIndices);
+			v.y = sin (p * (float)(AI_MATH_TWO_PI)/face.mNumIndices);
 		}
 	}
 }
